Deprecated

Options for computing EN estimates.

en_options_aug_lars(use_gram = c("auto", "yes", "no"), eps = 1e-12)

en_options_dal(
  maxit = 100,
  eps = 1e-08,
  eta_mult = 2,
  eta_start_numerator = 0.01,
  eta_start,
  preconditioner = c("approx", "none", "diagonal"),
  verbosity = 0
)

Arguments

use_gram
ignored. Should the Gram matrix be pre-computed.

eps
ignored. Numeric tolerance for convergence.

maxit
maximum number of iterations allowed.

eta_mult
multiplier to increase eta at each iteration.

eta_start_numerator
if eta_start is missing, it is defined by eta_start = eta_start_numerator / lambda.

eta_start
ignored. The start value for eta.

preconditioner
ignored. Preconditioner for the numerical solver. If none, a standard solver will be used, otherwise the faster preconditioned conjugate gradient is used.

verbosity
ignored.

Functions

en_options_aug_lars: Superseded by en_lars_options().

en_options_dal: Superseded by en_dal_options()

Warning

Do not use these functions in new code. They may be removed from future versions of the package.

See also