Use Coordinate Descent to Solve Elastic Net Problems

Use Coordinate Descent to Solve Elastic Net Problems

Usage

en_cd_options(max_it = 1000, reset_it = 8)

Arguments

max_it

maximum number of iterations.

reset_it

number of iterations after which the residuals are re-computed from scratch, to prevent numerical drifts from incremental updates.

See also