Deprecated

Options for computing initial estimates via ENPY. Superseded by enpy_options().

initest_options(
  keep_solutions = 5,
  psc_method = c("exact", "rr"),
  maxit = 10,
  maxit_pense_refinement = 5,
  eps = 1e-06,
  psc_keep = 0.5,
  resid_keep_method = c("proportion", "threshold"),
  resid_keep_prop = 0.6,
  resid_keep_thresh = 2,
  mscale_eps = 1e-08,
  mscale_maxit = 200
)

Arguments

keep_solutions
how many initial estimates should be kept to perform full PENSE iterations?

psc_method
The method to use for computing the principal sensitivity components. See details for the possible choices.

maxit
maximum number of refinement iterations.

maxit_pense_refinement
ignored. Maximum number of PENSE iterations to refine initial estimator.

eps
ignored. Numeric tolerance for convergence.

psc_keep
proportion of observations to keep based on the PSC scores.

resid_keep_method
How to clean the data based on large residuals. If "proportion", observations with the smallest resid_keep_prop residuals will be retained. If "threshold", all observations with scaled residuals smaller than the threshold resid_keep_thresh will be retained.

resid_keep_prop, resid_keep_thresh
proportion or threshold for observations to keep based on their residual.

mscale_eps, mscale_maxit
ignored. Maximum number of iterations and numeric tolerance for the M-scale.

Warning

Do not use this function in new code. It may be removed from future versions of the package.

See also