Deprecated

Additional options for computing penalized EN MM-estimates. Superseded by mm_algorithm_options() and options supplied directly to pensem_cv().

mstep_options(
  cc = 3.44,
  maxit = 1000,
  eps = 1e-06,
  adjust_bdp = FALSE,
  verbosity = 0,
  en_correction = TRUE
)

Arguments

cc

ignored. Tuning constant for the M-estimator.

maxit

maximum number of iterations allowed.

eps

ignored. Numeric tolerance for convergence.

adjust_bdp

ignored. Should the breakdown point be adjusted based on the effective degrees of freedom?

verbosity

ignored. Verbosity of the algorithm.

en_correction

ignored. Should the corrected EN estimator be used to choose the optimal lambda with CV. If TRUE, as by default, the estimator is "bias corrected".

Warning

Do not use this function in new code. It may be removed from future versions of the package.

See also