Extract Residuals

Extract residuals from a PENSE (or LS-EN) regularization path fitted by pense() or elnet().

# S3 method for pense_fit
residuals(object, lambda, exact = deprecated(), correction = deprecated(), ...)

Arguments

object
PENSE regularization path to extract residuals from.

lambda
a single value of the penalty parameter.

exact
defunct Always gives a warning if lambda is not part of the fitted sequence and coefficients need to be interpolated.

correction
defunct.

...
currently not used.

Value

a numeric vector of residuals for the given penalization level.

See also

Examples

# Compute the LS-EN regularization path for Freeny's revenue data
# (see ?freeny)
data(freeny)
x <- as.matrix(freeny[ , 2:5])

regpath <- elnet(x, freeny$y, alpha = 0.75)

# Predict the response using a specific penalization level
predict(regpath, newdata = freeny[1:5, 2:5],
        lambda = regpath$lambda[10])

#>  1962.25   1962.5  1962.75     1963  1963.25 
#> 9.071638 9.075877 9.082341 9.091051 9.103643 

# Extract the residuals at a certain penalization level
residuals(regpath, lambda = regpath$lambda[5])

#>              Qtr1         Qtr2         Qtr3         Qtr4
#> 1962              -0.396169224 -0.398919454 -0.378220695
#> 1963 -0.384593892 -0.294338479 -0.278030430 -0.259140970
#> 1964 -0.265543684 -0.219857551 -0.206023748 -0.173545125
#> 1965 -0.192086388 -0.146253918 -0.133786817 -0.096671191
#> 1966 -0.090347926 -0.044601264 -0.027307275 -0.015531362
#> 1967  0.006739697  0.037235389  0.038869333  0.071407419
#> 1968  0.086275062  0.102036759  0.141643505  0.167310507
#> 1969  0.175437660  0.211500792  0.222942549  0.260029329
#> 1970  0.247870647  0.296181577  0.294470169  0.278217170
#> 1971  0.306686546  0.334684047  0.353729154  0.367702082

# Select penalization level via cross-validation
cv_results <- elnet_cv(x, freeny$y, alpha = 0.5, cv_repl = 10,
                       cv_k = 4)

# Predict the response using the "best" penalization level
predict(cv_results, newdata = freeny[1:5, 2:5])

#>  1962.25   1962.5  1962.75     1963  1963.25 
#> 8.794094 8.806179 8.823862 8.841466 8.882830 

# Extract the residuals at the "best" penalization level
residuals(cv_results)^2

#>              Qtr1         Qtr2         Qtr3         Qtr4
#> 1962              3.005095e-06 2.193146e-04 8.104134e-05
#> 1963 8.097514e-04 6.090920e-04 1.791670e-04 1.963012e-04
#> 1964 1.853747e-04 4.862530e-05 1.977086e-06 1.739159e-04
#> 1965 6.665374e-04 1.405364e-05 1.042364e-04 6.203151e-06
#> 1966 2.193333e-04 2.245208e-04 8.200783e-05 9.905717e-06
#> 1967 6.170454e-06 1.283443e-04 2.252591e-05 6.226242e-06
#> 1968 2.232229e-04 2.683152e-04 1.890132e-05 8.440375e-05
#> 1969 7.791054e-05 6.196729e-04 1.163387e-04 7.753244e-04
#> 1970 4.981744e-04 2.140198e-04 4.137242e-05 3.568147e-04
#> 1971 1.469863e-04 2.686597e-05 9.557948e-06 5.428973e-05

# Extract the residuals at a more parsimonious penalization level
residuals(cv_results, lambda = 'se')

#>               Qtr1          Qtr2          Qtr3          Qtr4
#> 1962               -0.0088984307 -0.0211159534 -0.0142274424
#> 1963 -0.0337770122  0.0223283796  0.0092186714  0.0098520440
#> 1964 -0.0174315975  0.0047486697 -0.0041022871  0.0107321308
#> 1965 -0.0284252350 -0.0043510109 -0.0117313263 -0.0029809124
#> 1966 -0.0158717550  0.0143460888  0.0079295164  0.0020844029
#> 1967 -0.0022608835  0.0114683449 -0.0048433907 -0.0007496397
#> 1968 -0.0129163919 -0.0141304701  0.0071741898  0.0115220813
#> 1969  0.0101239940  0.0269147816  0.0128947265  0.0304762305
#> 1970 -0.0185378448  0.0192234266 -0.0024645664 -0.0154773994
#> 1971 -0.0069036801  0.0100063851  0.0081723476  0.0119808180