Predict Method for PENSE Fits

Predict response values using a PENSE (or LS-EN) regularization path fitted by pense() or elnet().

# S3 method for pense_fit
predict(
  object,
  newdata,
  lambda,
  exact = deprecated(),
  correction = deprecated(),
  ...
)

Arguments

object
PENSE regularization path to extract residuals from.

newdata
an optional matrix of new predictor values. If missing, the fitted values are computed.

lambda
a single value of the penalty parameter.

exact
defunct Always gives a warning if lambda is not part of the fitted sequence and coefficients need to be interpolated.

correction
defunct.

...
currently not used.

Value

a numeric vector of residuals for the given penalization level.

See also

Examples

# Compute the LS-EN regularization path for Freeny's revenue data
# (see ?freeny)
data(freeny)
x <- as.matrix(freeny[ , 2:5])

regpath <- elnet(x, freeny$y, alpha = 0.75)

# Predict the response using a specific penalization level
predict(regpath, newdata = freeny[1:5, 2:5],
        lambda = regpath$lambda[10])

#>  1962.25   1962.5  1962.75     1963  1963.25 
#> 9.071638 9.075877 9.082341 9.091051 9.103643 

# Extract the residuals at a certain penalization level
residuals(regpath, lambda = regpath$lambda[5])

#>              Qtr1         Qtr2         Qtr3         Qtr4
#> 1962              -0.396169224 -0.398919454 -0.378220695
#> 1963 -0.384593892 -0.294338479 -0.278030430 -0.259140970
#> 1964 -0.265543684 -0.219857551 -0.206023748 -0.173545125
#> 1965 -0.192086388 -0.146253918 -0.133786817 -0.096671191
#> 1966 -0.090347926 -0.044601264 -0.027307275 -0.015531362
#> 1967  0.006739697  0.037235389  0.038869333  0.071407419
#> 1968  0.086275062  0.102036759  0.141643505  0.167310507
#> 1969  0.175437660  0.211500792  0.222942549  0.260029329
#> 1970  0.247870647  0.296181577  0.294470169  0.278217170
#> 1971  0.306686546  0.334684047  0.353729154  0.367702082

# Select penalization level via cross-validation
cv_results <- elnet_cv(x, freeny$y, alpha = 0.5, cv_repl = 10,
                       cv_k = 4)

# Predict the response using the "best" penalization level
predict(cv_results, newdata = freeny[1:5, 2:5])

#>  1962.25   1962.5  1962.75     1963  1963.25 
#> 8.795162 8.807070 8.824535 8.842175 8.882970 

# Extract the residuals at the "best" penalization level
residuals(cv_results)^2

#>              Qtr1         Qtr2         Qtr3         Qtr4
#> 1962              7.848897e-06 2.464835e-04 9.359960e-05
#> 1963 8.505988e-04 6.021996e-04 1.637041e-04 1.795771e-04
#> 1964 2.010055e-04 4.526007e-05 3.191828e-06 1.650531e-04
#> 1965 6.871886e-04 1.397538e-05 1.084591e-04 6.238417e-06
#> 1966 2.235974e-04 2.227127e-04 7.845685e-05 8.735269e-06
#> 1967 5.946661e-06 1.285224e-04 2.320350e-05 4.855345e-06
#> 1968 2.138126e-04 2.572342e-04 2.288132e-05 9.043482e-05
#> 1969 7.984318e-05 6.318114e-04 1.211829e-04 7.936170e-04
#> 1970 4.757803e-04 2.341024e-04 3.523648e-05 3.416734e-04
#> 1971 1.285713e-04 3.422934e-05 1.431594e-05 6.324661e-05

# Extract the residuals at a more parsimonious penalization level
residuals(cv_results, lambda = 'se')

#>              Qtr1         Qtr2         Qtr3         Qtr4
#> 1962              -0.010525453 -0.022619883 -0.015569511
#> 1963 -0.035107544  0.021470216  0.008204877  0.008871462
#> 1964 -0.018329498  0.004095609 -0.004768904  0.010118697
#> 1965 -0.029022279 -0.004652820 -0.012121834 -0.003197425
#> 1966 -0.016137204  0.014152676  0.007720194  0.001900761
#> 1967 -0.002223898  0.011524214 -0.004779388 -0.000397902
#> 1968 -0.012479258 -0.013629057  0.007790677  0.012095335
#> 1969  0.010559764  0.027477195  0.013522814  0.031191235
#> 1970 -0.017586425  0.020296752 -0.001419612 -0.014534776
#> 1971 -0.005707253  0.011190639  0.009432874  0.013193935